Create page participant list export
