+++ src/main/resources/templates/ranking.html
@@ -59,7 +59,7 @@
                 <ul class="multi">
                     <li class="col-lg-12">
                         <input class="select-element" type="checkbox" name="all" id="all">
-                        <label for="all">all</label>
+                        <label for="all">alle</label>
 
                         <!-- disciplines -->
                         <ul class="row">
